Move mod-level ActorInits out of the engine.
