+++ OpenRA.Mods.TS/UtilityCommands/ImportTSMapCommand.cs
@@ -15,9 +15,9 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using OpenRA.FileSystem;
+using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.FileFormats;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Traits;
 
 namespace OpenRA.Mods.TS.UtilityCommands
 {
